Semi working version. Only supports single deep chunks for now.
Also, loading is still super slow, was hoping this way would be faster. But no :( Going to try out a middle-way between this concept and the previous one. Still compressing similar to the old way, but with a separate version for land/water, and a lower res one for colours/types.